combined functions files into one (m4_functions)
