+++ dash/M4/m4_app.py
@@ -19,8 +19,7 @@ from dash.dependencies import Input, Output
 import plotly.express as px
 import plotly.graph_objects as go
 
-import mortgage_funcs # source file for mortgage functions
-import stock_funcs # source file for stock functions
+import m4_functions
 import m4_parameters 
 
 
@@ -30,8 +29,8 @@ def custom_date_parser(date):
     return pd.datetime.strptime(date, "%Y-%m-%d") 
 
 #### Fetch data 
-mt, mt_summary = mortgage_funcs.mt_fetch()
-st, st_summary = stock_funcs.st_fetch()
+mt, mt_summary = m4_functions.mt_fetch()
+st, st_summary = m4_functions.st_fetch()
 
 #### graphical elements ####
 
@@ -41,26 +40,27 @@ mt_interest = px.scatter(mt, x="date", y=["prin_total", "int_total"])
 mt_interest.update_layout(hovermode='x')
 
 # adjust plots for time of day
-mt_balance = mortgage_funcs.time_of_day(mt_balance)
-mt_interest = mortgage_funcs.time_of_day(mt_interest)
+mt_balance = m4_functions.time_of_day(mt_balance)
+mt_interest = m4_functions.time_of_day(mt_interest)
 
 # set up tables
-mt_table = mortgage_funcs.table_setup(mt)
-mt_summary_table = mortgage_funcs.table_setup(mt_summary)
-st_table = mortgage_funcs.table_setup(st)
-st_summary_table = mortgage_funcs.table_setup(st_summary)
+mt_table = m4_functions.table_setup(mt)
+mt_summary_table = m4_functions.table_setup(mt_summary)
+st_table = m4_functions.table_setup(st)
+st_summary_table = m4_functions.table_setup(st_summary)
 
 #### app layout ####
 
-app.layout = html.Div(style={'backgroundColor': mortgage_funcs.colors['background']}, children=[
+app.layout = html.Div(style={'backgroundColor': m4_functions.colors['background']}, children=[
     html.H3(
         children="Marc's Money-Making Machine",
         style={'textAlign': 'center','color': '#2fa4e7'}
     ),
     dcc.Tabs(id='tabs-example', value='tab-1', children=[
-        dcc.Tab(label='Mortgage plots', value='tab-1', style=m4_parameters.tab_style, selected_style=m4_parameters.tab_selected_style),
+        dcc.Tab(label='Mortgage charts', value='tab-1', style=m4_parameters.tab_style, selected_style=m4_parameters.tab_selected_style),
         dcc.Tab(label='Mortgage table', value='tab-2', style=m4_parameters.tab_style, selected_style=m4_parameters.tab_selected_style),
         dcc.Tab(label='Stock table', value='tab-3', style=m4_parameters.tab_style, selected_style=m4_parameters.tab_selected_style),
+        dcc.Tab(label='Stock charts', value='tab-4', style=m4_parameters.tab_style, selected_style=m4_parameters.tab_selected_style),
     ]),
     html.Div(id='tabs-example-content')
 ])
@@ -119,6 +119,27 @@ def render_content(tab):
         st_table
         ]),  
     )
+
+    if tab == 'tab-4':
+        return (html.Div([
+        html.H3(children='Balance',
+        style={'textAlign': 'center','color': '#2fa4e7'}),
+
+        dcc.Graph(
+            id='graph1',
+            figure=mt_balance
+        ),  
+    ]),
+    # New Div for all elements in the new 'row' of the page
+    html.Div([
+        html.H3(children='Principal & Interest',
+        style={'textAlign': 'center','color': '#2fa4e7'}),
+
+        dcc.Graph(
+            id='graph2',
+            figure=mt_interest
+        ),  
+    ]))
 
 if __name__ == '__main__':
     app.run_server(debug=True)
